Add Remark Command
There is no command to add a remark to each contact.

User might want to add some notes about a contact and adding a remark address this.

Add a new remark command that allows users to add/remove notes on a contact.

Additional tests have been written to test the functionality of the new RemarkCommand.